experimental support for reactive Jakarta Data repositories
some cleanup
